+++ src/@classes/public/IsometricGroup/index.ts
@@ -1,0 +1,2 @@
+export { IsometricGroup } from './IsometricGroup';
+export { IsometricGroupProps } from './types';