+++ src/@classes/public/IsometricGroup/types.ts
@@ -1,0 +1,5 @@
+export interface IsometricGroupProps {
+    right?: number;
+    left?: number;
+    top?: number;
+}